Factor out proxies into their own file. Rewrite to verify that fewer files are loaded
